requirements/mysql.txt
Flask
SQLAlchemy
alembic
Flask-SQLAlchemy
wtforms
Flask-WTF
Flask-Admin
pytz==2013b
Flask-BabelEx
Flask-Login
Flask-Principal
thrift
blinker
selenium
#twisted
mysql-python
dbf
patool
fabric
lxml
dateutils

-r ../pysimplelogs/requirements.txt